씨부레
강사님 하라는 거
0. 충돌 델타 타임 충돌

내가 하라는 거
1. 타일 정밀 충돌

+++ KCore/KSprite_Render.cpp
@@ -68,7 +68,7 @@ void KSprite_Render::set_idx(const int& _Idx)
 	}
 	else if((int)VectorMySplit.size() <= _Idx)
 	{
-		idx = VectorMySplit.size() - 1;
+		idx = (int)VectorMySplit.size() - 1;
 	}
 	else
 	{
